requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
--index-url https://read-only:z4WIE57y889O93aX@pypicloud.canvasmedical.com/simple/

-e git+https://github.com/django/asgiref.git@daebab19124cf85150fff5ffa2457beee92cf49f#egg=asgiref
-e git+https://github.com/django/channels.git@d6643d96dd59e153cda13681a6d5a8963b9d577c#egg=channels
aioredis==1.0.0           # via channels-redis
appnope==0.1.0            # via ipython
arrow==0.12.1
asn1crypto==0.24.0        # via cryptography
async-timeout==2.0.0      # via aioredis
attrs==17.4.0             # via automat, service-identity
autobahn==17.10.1         # via daphne
automat==0.6.0            # via twisted
canvas-common==2.1.1
certifi==2018.1.18        # via requests
cffi==1.11.4              # via cryptography
channels-redis==2.1.0
chardet==3.0.4            # via requests
configparser==3.5.0       # via tini
constantly==15.1.0        # via twisted
cryptography==2.1.4
daphne==2.0.4
decorator==4.2.1          # via ipython, traitlets
dj-database-url==0.4.2
django-cors-headers==2.1.0
django-extensions==1.9.9
django-filter==1.1.0
django==2.0.2
djangorestframework==3.7.7
env_tools==2.2.0
h2==3.0.1                 # via twisted
hiredis==0.2.0            # via aioredis
hpack==3.0.0              # via h2
hyperframe==5.1.0         # via h2
hyperlink==17.3.1         # via twisted
idna==2.6                 # via cryptography, requests, twisted
incremental==17.5.0       # via twisted
ipython-genutils==0.2.0   # via traitlets
ipython==6.2.1
jedi==0.11.1              # via ipython
msgpack==0.5.6            # via channels-redis
parso==0.1.1              # via jedi
pexpect==4.4.0            # via ipython
pickleshare==0.7.4        # via ipython
priority==1.3.0           # via twisted
prompt-toolkit==1.0.15    # via ipython
psutil==5.4.3
psycopg2==2.7.4
ptyprocess==0.5.2         # via pexpect
pyasn1-modules==0.2.1     # via service-identity
pyasn1==0.4.2             # via pyasn1-modules, service-identity
pycparser==2.18           # via cffi
pygments==2.2.0           # via ipython
pyjwt==1.5.3              # via canvas-common
pyopenssl==17.5.0         # via service-identity, twisted
python-dateutil==2.6.1    # via arrow
pytz==2018.3              # via django
raven==6.5.0
requests==2.18.4
service-identity==17.0.0  # via twisted
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via autobahn, automat, cryptography, django-extensions, env-tools, prompt-toolkit, pyopenssl, python-dateutil, tini, traitlets, txaio
stringcase==1.2.0
tini==3.0.1               # via env-tools
traitlets==4.3.2          # via ipython
twisted[http2,tls]==17.9.0
txaio==2.8.2              # via autobahn
typing==3.6.4             # via django-extensions
urllib3==1.22             # via requests
uwsgi==2.0.16
wcwidth==0.1.7            # via prompt-toolkit
whitenoise==3.3.1
zope.interface==4.4.3     # via twisted
